+++ apps/console/src/app/app.tsx
@@ -13,7 +13,6 @@ import { DarkModeEnabler, Layout } from '@qovery/pages/layout'
 import { PageLogin, PageLogoutFeature } from '@qovery/pages/login'
 import { AssistantContext } from '@qovery/shared/assistant/feature'
 import { useAuth, useInviteMember } from '@qovery/shared/auth'
-import { ProtectedRoute } from '@qovery/shared/router'
 import { HELM_DEFAULT_VALUES, KUBECONFIG, LOGIN_URL, LOGOUT_URL, PREVIEW_CODE } from '@qovery/shared/routes'
 import { routeTree } from '@qovery/shared/routes'
 import { LoadingScreen } from '@qovery/shared/ui'
@@ -22,6 +21,7 @@ import { GTM, LOGROCKET, NODE_ENV, NX_GIT_SHA, QOVERY_API } from '@qovery/shared
 import { useAuthInterceptor } from '@qovery/shared/utils'
 import { queryClient } from '../query-client'
 import PreviewCode from './components/preview-code'
+import { ProtectedRoute } from './components/protected-route'
 import ScrollToTop from './components/scroll-to-top'
 import { ROUTER } from './router/main.router'
