+++ apps/console/src/app/components/protected-route.spec.tsx
@@ -1,5 +1,5 @@
 import { render } from '__tests__/utils/setup-jest'
-import ProtectedRoute from '../protected-route/protected-route'
+import ProtectedRoute from './protected-route'
 
 describe('ProtectedRoute', () => {
   const ProtectedComponent = () => {
