+++ apps/console/src/app/components/protected-route.tsx
